refactor(ch168): move mutations to own dir for client
