+++ client/src/__mutations__/DeleteRecord.gql.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const DELETE_RECORD = gql`
+  mutation DeleteRecord($id: ID!) {
+    deleteRecord(id: $id) {
+      id
+    }
+  }
+`;
