+++ client/src/__mutations__/InsertRecord.gql.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const INSERT_RECORD = gql`
+  mutation InsertRecord($input: InsertDatasetRecordInput!) {
+    insertRecord(input: $input) {
+      id
+    }
+  }
+`;
